+++ src/main/java/christmas/domain/menu/MenuType.java
@@ -1,4 +1,4 @@
-package christmas.domain.order.menu;
+package christmas.domain.menu;
 
 public enum MenuType {
