+++ src/main/java/christmas/domain/order/OrderLine.java
@@ -1,6 +1,6 @@
 package christmas.domain.order;
 
-import christmas.domain.order.menu.Menu;
+import christmas.domain.menu.Menu;
 
 public class OrderLine {
