+++ src/test/java/christmas/domain/order/OrderManagerTest.java
@@ -4,10 +4,10 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.assertj.core.api.Assertions.tuple;
 
-import christmas.domain.order.menu.Menu;
-import christmas.domain.order.menu.MenuBoard;
-import christmas.domain.order.menu.MenuName;
-import christmas.domain.order.menu.MenuType;
+import christmas.domain.menu.Menu;
+import christmas.domain.menu.MenuBoard;
+import christmas.domain.menu.MenuName;
+import christmas.domain.menu.MenuType;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
